atoc-data-pipeline/requirements.txt
# AtoC Korea 데이터 파이프라인
requests>=2.28.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
google-genai>=1.0.0
supabase>=2.0.0
h3>=3.7.0
